// resources/js/components/forms-module/types/types.ts
export interface DocumentType {
  id: number;
  name: string;
  slug: string;
  description?: string;
}